renamed model to objects and added file heirarchy
