+++ .gradle/config.properties
@@ -1,2 +1,2 @@
-#Mon Jan 08 00:49:48 CST 2024
-java.home=C\:\\Program Files\\Android\\Android Studio\\jbr
+#Sun Jan 14 23:26:24 CST 2024
+java.home=M\:\\Android\\jbr
