+++ app/src/main/java/com/example/a431transit/app/MainActivity.java
@@ -9,7 +9,7 @@ import android.annotation.SuppressLint;
 import android.os.Bundle;
 
 import com.example.a431transit.R;
-import com.example.a431transit.model.stops.BusStop;
+import com.example.a431transit.objects.stops.BusStop;
 import com.example.a431transit.util.api_communication.TransitAPIClient;
 import com.example.a431transit.util.api_communication.TransitAPIService;
 import com.example.a431transit.databinding.ActivityMainBinding;
