--- app/src/main/java/com/example/a431transit/model/stops/Street.java
@@ -1,32 +1,0 @@
-package com.example.a431transit.model.stops;
-
-import com.google.gson.annotations.SerializedName;
-
-public class Street {
-    @SerializedName("key")
-    private int key;
-
-    @SerializedName("name")
-    private String name;
-
-    @SerializedName("type")
-    private String type;
-
-    public Street(int k, String n, String t) {
-        key = k;
-        name = n;
-        type = t;
-    }
-
-    public int getKey() {
-        return key;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public String getType() {
-        return type;
-    }
-}
