+++ app/src/main/java/com/example/a431transit/objects/arrivals/route_schedules/scheduled_stops/bus/Bus.java
@@ -1,0 +1,27 @@
+package com.example.a431transit.objects.arrivals.route_schedules.scheduled_stops.bus;
+
+import com.google.gson.annotations.SerializedName;
+
+public class Bus {
+
+    @SerializedName("key")
+    private int key;
+
+    @SerializedName("bike-rack")
+    private String bikeRack;
+
+    @SerializedName("wifi")
+    private String wifi;
+
+    public int getKey() {
+        return key;
+    }
+
+    public String getBikeRack() {
+        return bikeRack;
+    }
+
+    public String getWifi() {
+        return wifi;
+    }
+}
