+++ app/src/main/java/com/example/a431transit/objects/arrivals/route_schedules/scheduled_stops/variant/Variant.java
@@ -1,0 +1,20 @@
+package com.example.a431transit.objects.arrivals.route_schedules.scheduled_stops.variant;
+
+import com.google.gson.annotations.SerializedName;
+
+public class Variant {
+
+    @SerializedName("key")
+    private String key;
+
+    @SerializedName("name")
+    private String name;
+
+    public String getKey() {
+        return key;
+    }
+
+    public String getName() {
+        return name;
+    }
+}
