+++ app/src/main/java/com/example/a431transit/objects/stops/centre/utm/UTM.java
@@ -1,0 +1,33 @@
+package com.example.a431transit.objects.stops.centre.utm;
+
+import com.google.gson.annotations.SerializedName;
+
+public class UTM {
+    @SerializedName("zone")
+    private String zone;
+
+    @SerializedName("x")
+    private int x;
+
+    @SerializedName("y")
+    private int y;
+
+    public UTM(String z, int x, int y) {
+        zone = z;
+        this.x = x;
+        this.y = y;
+    }
+
+    public String getZone() {
+        return zone;
+    }
+
+    public int getX() {
+        return x;
+    }
+
+    public int getY() {
+        return y;
+    }
+
+}
