+++ app/src/main/java/com/example/a431transit/util/bus_stop_expandable_list/BusStopGridAdapter.java
@@ -9,7 +9,7 @@ import android.widget.ImageView;
 import android.widget.TextView;
 
 import com.example.a431transit.R;
-import com.example.a431transit.model.stops.BusStop;
+import com.example.a431transit.objects.stops.BusStop;
 import com.example.a431transit.util.api_communication.TransitAPIService;
 
 import java.util.List;
