+++ app/src/main/java/com/example/a431transit/util/bus_stop_expandable_list/BusStopGridViewItemClickInterface.java
@@ -1,6 +1,6 @@
 package com.example.a431transit.util.bus_stop_expandable_list;
 
-import com.example.a431transit.model.stops.BusStop;
+import com.example.a431transit.objects.stops.BusStop;
 
 public interface BusStopGridViewItemClickInterface {
     void onGridViewBusStopClick(BusStop busStop);
