+++ app/src/main/java/com/example/a431transit/util/bus_stop_list/BusArrivalAdapter.java
@@ -10,7 +10,7 @@ import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
 
 import com.example.a431transit.R;
-import com.example.a431transit.model.arrivals.ArrivalInstance;
+import com.example.a431transit.objects.arrivals.ArrivalInstance;
 import com.example.a431transit.util.api_communication.TransitAPIService;
 
 import java.util.List;
@@ -69,7 +69,6 @@ public class BusArrivalAdapter extends RecyclerView.Adapter<BusArrivalAdapter.Bu
         else {
             return 0;
         }
-
     }
 
     @Override
