+++ app/src/main/java/com/example/a431transit/util/storage_managers/CategoriesManager.java
@@ -2,7 +2,7 @@ package com.example.a431transit.util.storage_managers;
 
 import android.content.Context;
 
-import com.example.a431transit.model.stops.BusStop;
+import com.example.a431transit.objects.stops.BusStop;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import com.google.gson.reflect.TypeToken;
