+++ local.properties
@@ -6,3 +6,5 @@
 # header note.
 #Sat Dec 30 22:13:53 CST 2023
 sdk.dir=C\:\\Users\\Leeroy Dilim\\AppData\\Local\\Android\\Sdk
+TRANSIT_API_KEY=KIsrRF-zLToUPFbSrHAl
+GOOGLE_API_KEY=AIzaSyDfo8pB8x7IDSu11hfrBbRltv_TpSBOSxc
