+++ secret.properties
@@ -1,0 +1,1 @@
+GOOGLE_API_KEY=AIzaSyDfo8pB8x7IDSu11hfrBbRltv_TpSBOSxc
